Basic Movement Added.
Basic Player movement. No gravity or ground collisions yet.

Basic Follower movement.
- trigger radius to "pick up" followers to initiate follow state.
- method to change state to waiting and stop following.
- method to Set build State to where, currently the follower agent will move towards the targeted grid cell.

Grid management.
The whole world is split into a grid and and followers are attached to a cell when they start "building". Unused cells are not referenced. When a used cell becomes unused it is removed from the used cell dictionary maintained by the gridManager.