Trigger for collisions with ants
When an object hits the collision trigger. The ant will switch states to ragdoll.

There should be a force applied, but it doesn't seem to be working correctly.
Objects are also stopped by the rigid bodies them selves. The object should continue moving through, maybe simply add more mass to the objects.