Ants have the ability to carry carryable objects.
The player can select a carriable object through state machines found in the Gamemanger. If the player "Mouse Hovers" over a carriable object and clicks. They may decide on a position to place it by clicking again. This will cause player grouped ants to run around the object and then "carry" it to the new location. The object itself is the actual thing moving. it has a navagent the is enabled when being carried and disabled when static.