Constraining carry object ant positions
By sampling the navmesh before returning the surrounding ant position. The returned position can be ensured to be on the nav mesh and the object still be "Picked Up".

Unwalkable navmesh regions must be added inside static navigation meshes. For example. The cubes in the test scene acidentally had nav mesh regions inside of them. Sampling the ant position would then cause the next available position to be inside the box mesh. A quick bandaid was to add an "Unwalkable" Quad inside the mesh.