Player ground checking fixed.
Player was not ground checking correctly and always increasing fall speed while appearing to be on the ground. This was due to the ground check capsule cast using the wrong calculations for the top and bottom offsets. The calculation now checks if the offset would be negative and adjusts to zero if it is.